camera work
Playing with camera here. Really liking the message but want it to make sense with the camera... trying to get it so that when space or enter are pushed to shoot, the perspective switches to the underside of the map, each half of the screen combining to make a whole.