Eyeball - what?
Ok, I got a little goofy here. I took the eye for an eye thing super litterally and tried to make a giant eye in the sky, sort of Katamari style. The thinking was that having the perspective switch with characters below and above ground could ba neat surreal juxtaposition, but... I couldn't figure out the prgramming, and tonally just a bit off. So going to scrap and try and focus on something more refined and in the vain of the last prototype for Monday's playtesting