Between loops interlinked
The solution to my programming woes of one mega loop? Make two! Not an elegant solution perhaps but it works...

Some of the things I have been thinking more about which I will elaborate more on in my final blog post is how the perspective switching I've chosen to use through the camera is really about drawing attention to the notion of omniscence ... who is being watched and controlled in this scene and why? Why are these tanks being forced to be reincarnated after destroying thier comrades? What kind of cruel tanks diety has created this universe!?

If an eye for an eye makes the whole world blind, and these tanks come form the same source, than really killing the other is like hurting it itself, no? And if this is the case - what kinds of thoughts would draw a tank to kill another tank?