Motivation
If an eye for an eye makes us blind...

What are the motivtions of tanks in TANKS! (?) A wife and child at home? Birth right? A cruel ancient diety that demands sacrifice?

I'd tried to do something more like (i.e., overwatching Katamari style God) in the version with the eyes, but here, hopefully, I've made something that's a bit less hammy and a bit more reflective.