Tanks
The two first experiments for this round are done. I'm glad I gave the first version pause when I got stuck on it; I'm much happier with the sensation of child's play that's happening in the second. As pippin pointed out, kids (the young ones at least) do that mash the toy into the ground thing.